Update and rename Button.py to pygame_utils.py
Note filename change.
Button.isClicked() is now Button.is_clicked()
Refactored code for consistency
